Project restructure to add the client refactor to Rust
