+++ Cargo.toml
@@ -1,19 +1,7 @@
-[package]
-name = "WinShowEQ"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-windows = { version = "0.62.2", features = [
-    "Win32_Foundation",
-    "Win32_Security",
-    "Win32_System_Diagnostics_Debug",
-    "Win32_System_Diagnostics_ToolHelp",
-    "Win32_System_Memory",
-    "Win32_System_Threading",
-    "Win32_System_WindowsProgramming",
-    "Win32_Networking_WinSock",
-    "Win32_System_Services",
-    "Win32_UI_Shell",
-    "Win32_UI_WindowsAndMessaging",
-] }
+[workspace]
+members = [
+    "common",
+    "server",
+    "client",
+]
+resolver = "2"
