+++ client/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "winshoweq-client"
+version = "0.1.0"
+edition = "2024"
+
+[[bin]]
+name = "mysEQ"
+path = "src/main.rs"
+
+[dependencies]
+common = { path = "../common" }
