+++ client/src/main.rs
@@ -1,0 +1,7 @@
+// WinShowEQ client — Rust replacement for MySEQ.exe (C# client).
+// See design/RUST_CLIENT_PLAN.md for the full milestone plan.
+// C1 milestone: project scaffold only.
+
+fn main() {
+    println!("WinShowEQ client — not yet implemented (see design/RUST_CLIENT_PLAN.md)");
+}
