+++ common/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2024"
