+++ server/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "winshoweq-server"
+version = "0.1.0"
+edition = "2024"
+
+[[bin]]
+name = "winshoweq"
+path = "src/main.rs"
+
+[dependencies]
+common = { path = "../common" }
+windows = { version = "0.62.2", features = [
+    "Win32_Foundation",
+    "Win32_Security",
+    "Win32_System_Diagnostics_Debug",
+    "Win32_System_Diagnostics_ToolHelp",
+    "Win32_System_Memory",
+    "Win32_System_Threading",
+    "Win32_System_WindowsProgramming",
+    "Win32_Networking_WinSock",
+    "Win32_System_Services",
+    "Win32_UI_Shell",
+    "Win32_UI_WindowsAndMessaging",
+] }
