+++ server/src/data/world.rs
@@ -1,0 +1,3 @@
+// WorldTime now lives in the common crate.
+// Re-exported here so server-internal modules keep `crate::data::world::WorldTime` working.
+pub use common::WorldTime;
