+++ server/src/mem_reader.rs
