+++ server/src/scanner.rs
